Use int for location components instead of uint8_t

+++ Engine/Location.h
@@ -5,7 +5,7 @@
 
 class Location {
  public:
-  Location(uint8_t x = 0, uint8_t y = 0) : x(x), y(y){};
+  Location(int x = 0, int y = 0) : x(x), y(y){};
   
   void add(Location const& delta) {
     x += delta.x;
@@ -14,16 +14,16 @@ class Location {
 
   Location const next(Direction dir) const noexcept {
     switch (dir) {
-      case Direction::UP:    return Location{x, (uint8_t)(y - 1)};
-      case Direction::DOWN:  return Location{x, (uint8_t)(y + 1)};
-      case Direction::LEFT:  return Location{(uint8_t)(x - 1), y};
-      case Direction::RIGHT: return Location{(uint8_t)(x + 1), y};
-      default:               return Location{x, y};
+      case Direction::UP:    return Location{x    , y - 1};
+      case Direction::DOWN:  return Location{x    , y + 1};
+      case Direction::LEFT:  return Location{x - 1, y    };
+      case Direction::RIGHT: return Location{x + 1, y    };
+      default:               return Location{x    , y    };
     };
   };
 
-  uint8_t x;
-  uint8_t y;
+  int x;
+  int y;
 };
 
 inline bool operator==(Location const& lhs, Location const& rhs) {
